+++ spec/support/shared/client_ack_examples.rb
@@ -4,6 +4,9 @@ shared_examples "supports client acks" do
       expect(subject.supports_client_acks?).to eq(true)
     end
   end
+
+  it { should respond_to :ack_message }
+  it { should respond_to :nack_message }
 end
 
 shared_examples "doesn't support client acks" do