+++ spec/units/message_driver/client_spec.rb
@@ -203,6 +203,68 @@ module MessageDriver
         end
       end
 
+      describe "#ack_message" do
+        let(:message) { double("message") }
+        let(:options) { {foo: :bar} }
+        before do
+          adapter_context.stub(:ack_message)
+        end
+        context "when the adapter supports client acks" do
+          before do
+            adapter_context.stub(:supports_client_acks?) { true }
+          end
+          it "calls #ack_message with the message" do
+            subject.ack_message(message)
+            adapter_context.should have_received(:ack_message).with(message, {})
+          end
+          it "passes the supplied options to ack_message" do
+            subject.ack_message(message, options)
+            adapter_context.should have_received(:ack_message).with(message, options)
+          end
+        end
+        context "when the adapter doesn't support client acks" do
+          before do
+            adapter_context.stub(:supports_client_acks?) { false }
+          end
+          it "doesn't call #ack_message" do
+            subject.ack_message(message)
+            adapter_context.should_not have_received(:ack_message)
+          end
+          it "logs a warning"
+        end
+      end
+
+      describe "#nack_message" do
+        let(:message) { double("message") }
+        let(:options) { {foo: :bar} }
+        before do
+          adapter_context.stub(:nack_message)
+        end
+        context "when the adapter supports client acks" do
+          before do
+            adapter_context.stub(:supports_client_acks?) { true }
+          end
+          it "calls #nack_message with the message" do
+            subject.nack_message(message)
+            adapter_context.should have_received(:nack_message).with(message, {})
+          end
+          it "passes the supplied options to nack_message" do
+            subject.nack_message(message, options)
+            adapter_context.should have_received(:nack_message).with(message, options)
+          end
+        end
+        context "when the adapter doesn't support client acks" do
+          before do
+            adapter_context.stub(:supports_client_acks?) { false }
+          end
+          it "doesn't call #ack_message" do
+            subject.nack_message(message)
+            adapter_context.should_not have_received(:nack_message)
+          end
+          it "logs a warning"
+        end
+      end
+
       describe "#subscribe" do
         let(:destination) { Broker.destination(:my_queue, "my_queue", exclusive: true) }
         let(:consumer_double) { lambda do |m| end }